Fix a potential overflow from large stacks
This was hitting in the tests.  It only matters on platforms where long
is 32 bits, so had been missed before.

+++ src/item.cpp
@@ -3038,7 +3038,7 @@ units::volume item::volume( bool integral ) const
     }
 
     if( count_by_charges() || made_of( LIQUID ) ) {
-        auto num = ret * charges;
+        auto num = ret * static_cast<int64_t>( charges );
         ret = num / type->stack_size;
         if( num % type->stack_size != 0_ml ) {
             ret += 1_ml;
